Add query context param `useDefaultExpressionExplain`
